Fix internal article query (#964)
* update abuse report subject template

* fix internal/article top month filters

* update resolution email spec with proper subject line

+++ app/controllers/internal/articles_controller.rb
@@ -17,7 +17,7 @@ class Internal::ArticlesController < Internal::ApplicationController
         per(50)
     when /top\-/
       @articles = Article.
-        where("published_at > ?", params[:state].split("-")[1].to_f.months.ago).
+        where("published_at > ?", params[:state].split("-")[1].to_i.months.ago).
         includes(:user).
         includes(:buffer_updates).
         order("positive_reactions_count DESC").
